+++ src/graphql/network.rs
@@ -10,7 +10,7 @@ use async_graphql::{
 use base64::Engine;
 use chrono::{DateTime, Utc};
 use giganto_client::ingest::network::{
-    Bootp, Conn, DceRpc, Dhcp, Dns, Ftp, Http, Kerberos, Ldap, MalformedDns, Mqtt, Nfs, Ntlm,
+    Bootp, Conn, DceRpc, Dhcp, Dns, Ftp, Http, Icmp, Kerberos, Ldap, MalformedDns, Mqtt, Nfs, Ntlm,
     Radius, Rdp, Smb, Smtp, Ssh, Tls,
 };
 #[cfg(feature = "cluster")]
@@ -33,25 +33,26 @@ use crate::graphql::client::{
     },
     derives::{
         BootpRawEvents, ConnRawEvents, DceRpcRawEvents, DhcpRawEvents, DnsRawEvents, FtpRawEvents,
-        HttpRawEvents, KerberosRawEvents, LdapRawEvents, MalformedDnsRawEvents, MqttRawEvents,
-        NetworkRawEvents as GraphQlNetworkRawEvents, NfsRawEvents, NtlmRawEvents, RadiusRawEvents,
-        RdpRawEvents, SearchBootpRawEvents, SearchConnRawEvents, SearchDceRpcRawEvents,
-        SearchDhcpRawEvents, SearchDnsRawEvents, SearchFtpRawEvents, SearchHttpRawEvents,
-        SearchKerberosRawEvents, SearchLdapRawEvents, SearchMalformedDnsRawEvents,
-        SearchMqttRawEvents, SearchNfsRawEvents, SearchNtlmRawEvents, SearchRadiusRawEvents,
-        SearchRdpRawEvents, SearchSmbRawEvents, SearchSmtpRawEvents, SearchSshRawEvents,
-        SearchTlsRawEvents, SmbRawEvents, SmtpRawEvents, SshRawEvents, TlsRawEvents,
-        bootp_raw_events, conn_raw_events, dce_rpc_raw_events, dhcp_raw_events, dns_raw_events,
-        ftp_raw_events, http_raw_events, kerberos_raw_events, ldap_raw_events,
-        malformed_dns_raw_events, mqtt_raw_events, network_raw_events, nfs_raw_events,
-        ntlm_raw_events, radius_raw_events, rdp_raw_events, search_bootp_raw_events,
-        search_conn_raw_events, search_dce_rpc_raw_events, search_dhcp_raw_events,
-        search_dns_raw_events, search_ftp_raw_events, search_http_raw_events,
-        search_kerberos_raw_events, search_ldap_raw_events, search_malformed_dns_raw_events,
-        search_mqtt_raw_events, search_nfs_raw_events, search_ntlm_raw_events,
-        search_radius_raw_events, search_rdp_raw_events, search_smb_raw_events,
-        search_smtp_raw_events, search_ssh_raw_events, search_tls_raw_events, smb_raw_events,
-        smtp_raw_events, ssh_raw_events, tls_raw_events,
+        HttpRawEvents, IcmpRawEvents, KerberosRawEvents, LdapRawEvents, MalformedDnsRawEvents,
+        MqttRawEvents, NetworkRawEvents as GraphQlNetworkRawEvents, NfsRawEvents, NtlmRawEvents,
+        RadiusRawEvents, RdpRawEvents, SearchBootpRawEvents, SearchConnRawEvents,
+        SearchDceRpcRawEvents, SearchDhcpRawEvents, SearchDnsRawEvents, SearchFtpRawEvents,
+        SearchHttpRawEvents, SearchIcmpRawEvents, SearchKerberosRawEvents, SearchLdapRawEvents,
+        SearchMalformedDnsRawEvents, SearchMqttRawEvents, SearchNfsRawEvents, SearchNtlmRawEvents,
+        SearchRadiusRawEvents, SearchRdpRawEvents, SearchSmbRawEvents, SearchSmtpRawEvents,
+        SearchSshRawEvents, SearchTlsRawEvents, SmbRawEvents, SmtpRawEvents, SshRawEvents,
+        TlsRawEvents, bootp_raw_events, conn_raw_events, dce_rpc_raw_events, dhcp_raw_events,
+        dns_raw_events, ftp_raw_events, http_raw_events, icmp_raw_events, kerberos_raw_events,
+        ldap_raw_events, malformed_dns_raw_events, mqtt_raw_events, network_raw_events,
+        nfs_raw_events, ntlm_raw_events, radius_raw_events, rdp_raw_events,
+        search_bootp_raw_events, search_conn_raw_events, search_dce_rpc_raw_events,
+        search_dhcp_raw_events, search_dns_raw_events, search_ftp_raw_events,
+        search_http_raw_events, search_icmp_raw_events, search_kerberos_raw_events,
+        search_ldap_raw_events, search_malformed_dns_raw_events, search_mqtt_raw_events,
+        search_nfs_raw_events, search_ntlm_raw_events, search_radius_raw_events,
+        search_rdp_raw_events, search_smb_raw_events, search_smtp_raw_events,
+        search_ssh_raw_events, search_tls_raw_events, smb_raw_events, smtp_raw_events,
+        ssh_raw_events, tls_raw_events,
     },
 };
 use crate::storage::{Database, FilteredIter, KeyExtractor};
@@ -1296,6 +1297,50 @@ struct RadiusRawEvent {
     message: String,
 }
 
+/// Represents an event extracted from the ICMP protocol.
+#[derive(SimpleObject, Debug)]
+#[cfg_attr(feature = "cluster", derive(ConvertGraphQLEdgesNode))]
+#[cfg_attr(feature = "cluster", graphql_client_type(names = [
+    icmp_raw_events::IcmpRawEventsIcmpRawEventsEdgesNode,
+    network_raw_events::NetworkRawEventsNetworkRawEventsEdgesNodeOnIcmpRawEvent
+]))]
+struct IcmpRawEvent {
+    /// Time the event started transmitting from a sensor
+    time: DateTime<Utc>,
+    /// Source IP Address
+    orig_addr: String,
+    /// Destination IP Address
+    resp_addr: String,
+    /// Protocol Number
+    proto: u8,
+    /// Start Time
+    start_time: DateTime<Utc>,
+    /// Duration
+    ///
+    /// It is measured in nanoseconds.
+    duration: StringNumberI64,
+    /// Packets Sent by Source
+    orig_pkts: StringNumberU64,
+    /// Packets Received by Destination
+    resp_pkts: StringNumberU64,
+    /// Layer 2 Bytes Sent by Source
+    orig_l2_bytes: StringNumberU64,
+    /// Layer 2 Bytes Received by Destination
+    resp_l2_bytes: StringNumberU64,
+    /// ICMP Type
+    icmp_type: u8,
+    /// ICMP Code
+    icmp_code: u8,
+    /// Identifier
+    id: u16,
+    /// Sequence Number
+    seq_num: u16,
+    /// Data Length
+    data_len: u16,
+    /// Payload
+    payload: Vec<u8>,
+}
+
 #[allow(clippy::enum_variant_names)]
 #[derive(Union)]
 enum NetworkRawEvents {
@@ -1318,6 +1363,7 @@ enum NetworkRawEvents {
     BootpRawEvent(BootpRawEvent),
     DhcpRawEvent(DhcpRawEvent),
     RadiusRawEvent(RadiusRawEvent),
+    IcmpRawEvent(IcmpRawEvent),
 }
 
 #[cfg(feature = "cluster")]
@@ -1381,6 +1427,9 @@ impl From<network_raw_events::NetworkRawEventsNetworkRawEventsEdgesNode> for Net
             network_raw_events::NetworkRawEventsNetworkRawEventsEdgesNode::RadiusRawEvent(
                 event,
             ) => NetworkRawEvents::RadiusRawEvent(event.into()),
+            network_raw_events::NetworkRawEventsNetworkRawEventsEdgesNode::IcmpRawEvent(event) => {
+                NetworkRawEvents::IcmpRawEvent(event.into())
+            }
         }
     }
 }
@@ -1647,6 +1696,29 @@ impl FromKeyValue<Radius> for RadiusRawEvent {
             nas_id: val.nas_id.clone(),
             nas_port_type: StringNumberU32(val.nas_port_type),
             message: val.message.clone(),
+        })
+    }
+}
+
+impl FromKeyValue<Icmp> for IcmpRawEvent {
+    fn from_key_value(key: &[u8], val: Icmp) -> Result<Self> {
+        Ok(IcmpRawEvent {
+            time: get_time_from_key(key)?,
+            orig_addr: val.orig_addr.to_string(),
+            resp_addr: val.resp_addr.to_string(),
+            proto: val.proto,
+            start_time: chrono::DateTime::from_timestamp_nanos(val.start_time),
+            duration: val.duration.into(),
+            orig_pkts: val.orig_pkts.into(),
+            resp_pkts: val.resp_pkts.into(),
+            orig_l2_bytes: val.orig_l2_bytes.into(),
+            resp_l2_bytes: val.resp_l2_bytes.into(),
+            icmp_type: val.icmp_type,
+            icmp_code: val.icmp_code,
+            id: val.id,
+            seq_num: val.seq_num,
+            data_len: val.data_len,
+            payload: val.payload,
         })
     }
 }
@@ -2064,6 +2136,20 @@ async fn handle_paged_radius_raw_events(
     handle_paged_events(store, filter, after, before, first, last).await
 }
 
+async fn handle_paged_icmp_raw_events(
+    ctx: &Context<'_>,
+    filter: NetworkFilter,
+    after: Option<String>,
+    before: Option<String>,
+    first: Option<i32>,
+    last: Option<i32>,
+) -> Result<Connection<String, IcmpRawEvent>> {
+    let db = ctx.data::<Database>()?;
+    let store = db.icmp_store()?;
+
+    handle_paged_events(store, filter, after, before, first, last).await
+}
+
 #[allow(clippy::too_many_lines)]
 async fn handle_network_raw_events(
     ctx: &Context<'_>,
@@ -2251,6 +2337,15 @@ async fn handle_network_raw_events(
                 last,
             )?;
 
+            let (icmp_iter, _) = get_peekable_iter(
+                &db.icmp_store()?,
+                &filter,
+                after.as_deref(),
+                before.as_deref(),
+                first,
+                last,
+            )?;
+
             let mut is_forward: bool = true;
             if before.is_some() || last.is_some() {
                 is_forward = false;
@@ -2276,6 +2371,7 @@ async fn handle_network_raw_events(
                 bootp_iter,
                 dhcp_iter,
                 radius_iter,
+                icmp_iter,
                 size,
                 is_forward,
             )
@@ -2793,6 +2889,32 @@ impl NetworkQuery {
             radius_raw_events::Variables,
             radius_raw_events::ResponseData,
             radius_raw_events
+        )
+    }
+
+    async fn icmp_raw_events(
+        &self,
+        ctx: &Context<'_>,
+        filter: NetworkFilter,
+        after: Option<String>,
+        before: Option<String>,
+        first: Option<i32>,
+        last: Option<i32>,
+    ) -> Result<Connection<String, IcmpRawEvent>> {
+        let handler = handle_paged_icmp_raw_events;
+        paged_events_in_cluster!(
+            ctx,
+            filter,
+            filter.sensor,
+            after,
+            before,
+            first,
+            last,
+            handler,
+            IcmpRawEvents,
+            icmp_raw_events::Variables,
+            icmp_raw_events::ResponseData,
+            icmp_raw_events
         )
     }
 
@@ -3346,6 +3468,34 @@ impl NetworkQuery {
             search_radius_raw_events
         )
     }
+
+    async fn search_icmp_raw_events(
+        &self,
+        ctx: &Context<'_>,
+        filter: SearchFilter,
+    ) -> Result<Vec<DateTime<Utc>>> {
+        let handler = |ctx: &Context<'_>, filter: &SearchFilter| {
+            let db = ctx.data::<Database>()?;
+            let store = db.icmp_store()?;
+            let exist_data = store
+                .batched_multi_get_from_ts(&filter.sensor, &filter.times)
+                .into_iter()
+                .collect::<BTreeSet<(DateTime<Utc>, Vec<u8>)>>();
+
+            Ok(collect_exist_times::<Icmp>(&exist_data, filter))
+        };
+
+        events_vec_in_cluster!(
+            ctx,
+            filter,
+            filter.sensor,
+            handler,
+            SearchIcmpRawEvents,
+            search_icmp_raw_events::Variables,
+            search_icmp_raw_events::ResponseData,
+            search_icmp_raw_events
+        )
+    }
 }
 
 #[allow(clippy::too_many_arguments, clippy::too_many_lines)]
@@ -3369,6 +3519,7 @@ fn network_connection(
     mut bootp_iter: Peekable<FilteredIter<Bootp>>,
     mut dhcp_iter: Peekable<FilteredIter<Dhcp>>,
     mut radius_iter: Peekable<FilteredIter<Radius>>,
+    mut icmp_iter: Peekable<FilteredIter<Icmp>>,
     size: usize,
     is_forward: bool,
 ) -> Result<Connection<String, NetworkRawEvents>> {
@@ -3397,6 +3548,7 @@ fn network_connection(
     let mut bootp_data = bootp_iter.next();
     let mut dhcp_data = dhcp_iter.next();
     let mut radius_data = radius_iter.next();
+    let mut icmp_data = icmp_iter.next();
 
     loop {
         let conn_ts = if let Some((ref key, _)) = conn_data {
@@ -3513,6 +3665,12 @@ fn network_connection(
             min_max_time(is_forward)
         };
 
+        let icmp_ts = if let Some((ref key, _)) = icmp_data {
+            get_time_from_key(key)?
+        } else {
+            min_max_time(is_forward)
+        };
+
         let selected = if is_forward {
             time.min(dns_ts)
                 .min(malformed_dns_ts)
@@ -3533,6 +3691,7 @@ fn network_connection(
                 .min(bootp_ts)
                 .min(dhcp_ts)
                 .min(radius_ts)
+                .min(icmp_ts)
         } else {
             time.max(dns_ts)
                 .max(malformed_dns_ts)
@@ -3553,6 +3712,7 @@ fn network_connection(
                 .max(bootp_ts)
                 .max(dhcp_ts)
                 .max(radius_ts)
+                .max(icmp_ts)
         };
 
         match selected {
@@ -3737,6 +3897,15 @@ fn network_connection(
                     radius_data = radius_iter.next();
                 }
             }
+            _ if selected == icmp_ts => {
+                if let Some((key, value)) = icmp_data {
+                    result_vec.push(Edge::new(
+                        base64_engine.encode(&key),
+                        NetworkRawEvents::IcmpRawEvent(IcmpRawEvent::from_key_value(&key, value)?),
+                    ));
+                    icmp_data = icmp_iter.next();
+                }
+            }
             _ => {}
         }
         if (result_vec.len() >= size)
@@ -3758,7 +3927,8 @@ fn network_connection(
                 && smtp_data.is_none()
                 && bootp_data.is_none()
                 && dhcp_data.is_none()
-                && radius_data.is_none())
+                && radius_data.is_none()
+                && icmp_data.is_none())
         {
             if conn_data.is_some()
                 || dns_data.is_some()
@@ -3779,6 +3949,7 @@ fn network_connection(
                 || bootp_data.is_some()
                 || dhcp_data.is_some()
                 || radius_data.is_some()
+                || icmp_data.is_some()
             {
                 has_next_value = true;
             }
@@ -3820,6 +3991,7 @@ impl_from_giganto_range_structs_for_graphql_client!(
     bootp_raw_events,
     dhcp_raw_events,
     radius_raw_events,
+    icmp_raw_events,
     search_conn_raw_events,
     search_dce_rpc_raw_events,
     search_dns_raw_events,
@@ -3838,7 +4010,8 @@ impl_from_giganto_range_structs_for_graphql_client!(
     search_tls_raw_events,
     search_bootp_raw_events,
     search_dhcp_raw_events,
-    search_radius_raw_events
+    search_radius_raw_events,
+    search_icmp_raw_events
 );
 
 #[cfg(feature = "cluster")]
@@ -3862,7 +4035,8 @@ impl_from_giganto_network_filter_for_graphql_client!(
     smb_raw_events,
     bootp_raw_events,
     dhcp_raw_events,
-    radius_raw_events
+    radius_raw_events,
+    icmp_raw_events
 );
 
 #[cfg(feature = "cluster")]
@@ -3885,5 +4059,6 @@ impl_from_giganto_search_filter_for_graphql_client!(
     search_tls_raw_events,
     search_bootp_raw_events,
     search_dhcp_raw_events,
-    search_radius_raw_events
+    search_radius_raw_events,
+    search_icmp_raw_events
 );